refactor: reorg concrete/lib
